+++ src/hooks/use-ga-event/index.ts
@@ -1,1 +1,1 @@
-export { default } from './useGaEvent';
+export { default } from './use-ga-event';
